update hit Transport
